Address few more comments from @das7pad
